# openapi.legacy.yaml
openapi: 3.0.3
info:
  title: Ceilgas
  version: 0.1.0
  description: >-
    Sound gas upper-bound analyser and certification API for Ethereum public functions.
  x-product-slug: ai-decentralized-ceilgas
  x-source: ai-in-decentralized+ai/research-paper_1811.10403v1/
servers:
  - url: https://api.example.com
...
paths:
...
  /v1/builds/{buildId}/functions:
    parameters:
      - name: buildId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: listFunctions
      tags: [Functions]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Functions page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListEnvelopePublicFunction" }
  /v1/functions/{functionId}:
    parameters:
      - name: functionId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getFunction
      tags: [Functions]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        "200":
          description: Public function
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DataEnvelopePublicFunction" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/recommendations:
    post:
      operationId: createLimitRecommendation
      tags: [Recommendations]
      security: [{ ApiKeyAuth: [] }, { BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LimitRecommendationCreate" }
      responses:
        "200":
          description: Recommendation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DataEnvelopeLimitRecommendation" }
        "400": { $ref: "#/components/responses/BadRequest" }
    get:
      operationId: listLimitRecommendations
      tags: [Recommendations]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Recommendations page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListEnvelopeLimitRecommendation" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
    PublicFunction:
      type: object
      required: [id, buildId, name]
      properties:
        id: { type: string }
        buildId: { type: string }
        name: { type: string }
        isConstantGas: { type: boolean }
        sizeMetrics:
          type: array
          items: { type: string }
...
    LimitRecommendation:
      type: object
      required: [id, functionId, recommendedGasLimit]
      properties:
        id: { type: string }
        functionId: { type: string }
        recommendedGasLimit: { type: integer }
        assumptions:
          type: object
          additionalProperties: { type: number }
        sound: { type: boolean }
    LimitRecommendationCreate:
      type: object
      required: [functionId, assumptions]
      properties:
        functionId: { type: string }
        assumptions:
          type: object
          additionalProperties: { type: number }
        callbackContext: { type: boolean, default: false }
...
    DataEnvelopePublicFunction:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/PublicFunction" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    ListEnvelopePublicFunction:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/PublicFunction" }
            nextCursor: { type: string }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
...
    DataEnvelopeLimitRecommendation:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/LimitRecommendation" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    ListEnvelopeLimitRecommendation:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/LimitRecommendation" }
            nextCursor: { type: string }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
...