# platform/api-server/src/domains/functions/handlers/function.handlers.ts
/**
 * Handlers for Function
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: functions
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { FunctionsDomainModule } from "../dependencies/functions-ddd.dependencies.js";
import type { GetFunctionInput, ListFunctionsInput } from "@ceilgas/services/functions";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/builds/{buildId}/functions
 * Operation: listFunctions
 */
export async function listFunctions(
  request: FastifyRequest<{ Params: { buildId: string }; Querystring: { cursor?: string | string[], limit?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: FunctionsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListFunctionsInput = {
    buildId: request.params.buildId,
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.functions.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for GET /v1/functions/{functionId}
 * Operation: getFunction
 */
export async function getFunction(
  request: FastifyRequest<{ Params: { functionId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: FunctionsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetFunctionInput = {
    functionId: request.params.functionId,
  };
  const result = await deps.useCases.functions.get.execute(input);
  return reply.code(200).send(result);
}