# platform/api-server/src/domains/recommendations/handlers/recommendation.handlers.ts
/**
 * Handlers for Recommendation
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: recommendations
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { RecommendationsDomainModule } from "../dependencies/recommendations-ddd.dependencies.js";
import type { CreateLimitRecommendationInput, ListLimitRecommendationsInput } from "@ceilgas/services/recommendations";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/recommendations
 * Operation: listLimitRecommendations
 */
export async function listLimitRecommendations(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], functionId?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: RecommendationsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListLimitRecommendationsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.recommendations.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/recommendations
 * Operation: createLimitRecommendation
 */
export async function createLimitRecommendation(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: RecommendationsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: CreateLimitRecommendationInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.recommendations.create.execute(input);
  return reply.code(200).send(result);
}